Fix request retries racing condition issue

+++ client/src/reducers/api/apiSlice.js
@@ -1,5 +1,7 @@
+/* eslint no-shadow: 0 */
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 import { setCredentials, resetCredentials } from '../authReducer'
+import { saveProfilePictureToDB } from '../thunk/authThunks'
 
 const BASE_URL = process.env.REACT_APP_BASE_URL
 const REFRESH_URI = process.env.REACT_APP_REFRESH_URI
@@ -22,31 +24,43 @@ const baseQuery = fetchBaseQuery({
  * so if request fails with an unauthorized error,
  * we can resend the request after getting a new token.
  */
+let refreshPromise = null // refresh lock using a Promise
+
 const baseQueryWithReAuth = async (args, api, extraOptions) => {
   let result = await baseQuery(args, api, extraOptions)
   // If token has expired, send a request to get a new token
   if (result?.error?.status === 403 || result?.error?.status === 401) {
-    console.log('Refresh token expired. Sending request for new token')
-    const refreshResult = await baseQuery(REFRESH_URI, api, extraOptions)
-    if (refreshResult?.data) {
-      const userState = api.getState().auth
-      api.dispatch(setCredentials({
-        ...userState,
-        ...refreshResult.data,
-      }))
-      // Retry original request with new token
+    if (!refreshPromise) {
+      refreshPromise = (async () => {
+        console.log('Refresh token expired. Sending request for new token')
+        const refreshResult = await baseQuery(REFRESH_URI, api, extraOptions)
+        if (refreshResult?.data) {
+          const userState = api.getState().auth
+          await api.dispatch(setCredentials({
+            ...userState,
+            ...refreshResult.data,
+          }))
+          return refreshResult.data
+        }
+        // If user is not authorized to get token, log out
+        await baseQuery(
+          {
+            url: REFRESH_URI,
+            method: 'DELETE',
+          },
+          api,
+          extraOptions,
+        )
+        await api.dispatch(saveProfilePictureToDB(null))
+        await api.dispatch(resetCredentials())
+        return null
+      })()
+    }
+    const newTokenData = await refreshPromise
+    refreshPromise = null
+    if (newTokenData) {
+      // Retry original request with the new token
       result = await baseQuery(args, api, extraOptions)
-      // If user is not authorized to get token, log out
-    } else {
-      await baseQuery(
-        {
-          url: REFRESH_URI,
-          method: 'DELETE',
-        },
-        api,
-        extraOptions,
-      )
-      api.dispatch(resetCredentials())
     }
   }
   return result
